requirements.txt
numpy
transformers[gpu]
Pillow
fastapi[all]